refactor: renamed dialoguer to prompt
