+++ src/hooks.rs
@@ -3,9 +3,9 @@ use std::io::{BufReader, Read, Write};
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 
-use crate::dialoguer::confirm;
 use crate::error::{Error, Result};
 use crate::ioutils::path_to_str;
+use crate::prompt::confirm;
 
 /// Structure representing data passed to hook scripts.
 ///
